linux-qcom-dtbbin.bbclass: Extend support to merge oot dt overlays
Extend the linux-qcom-dtbbin class to overlay out-of-tree DTBOs on top of
the base device-tree blobs produced by the kernel.

The list of DTBOs to be overlaid on top of the base kernel device-tree must
be specified in the DTB-specific KERNEL_DTB_OVERLAYS flag. For example,
to overlay my-dt-overlay.dtbo on top of kernel-dt.dtb, use the following:
KERNEL_DTB_OVERLAYS[kernel-dt] = "my-dt-overlay.dtbo"

If there are multiple DTBO files, they must be listed in dependency order.
That is, the DTBO file that depends on another DTBO file should be listed
later in the list. For example, if my-dtbo1.dtbo depends on my-dtbo2.dtbo,
they should be mentioned as:
KERNEL_DTB_OVERLAYS[kernel-dt] = "my-dtbo2.dtbo my-dtbo1.dtbo"

Recipes generating out-of-tree DTBOs must be listed in KERNEL_DTBO_PROVIDERS
for BitBake to set the correct task dependency order and ensure DTBOs are built
and available for overlaying. For example, if the 'my-dtb.bb' recipe generates
'my-dt-overlay.dtbo', it should be mentioned as:
KERNEL_DTBO_PROVIDERS = "my-dtb"

Signed-off-by: Viswanath Kraleti <quic_vkraleti@quicinc.com>